+++ packages/adminui-vue/components/controls/UCodeMirror.vue
@@ -55,6 +55,13 @@ module.exports = {
     editorMode: {
       type: String,
       default: 'application/x-javascript'
+    },
+    /**
+     * Optional function what return a hints. See hint/show-hint.js section in https://codemirror.net/doc/manual.html#addons
+     * Called with one parameter - codeMirror instance
+     */
+    hintsFunction: {
+      type: Function
     }
   },
 
@@ -143,47 +150,9 @@ module.exports = {
     },
 
     showTemplates () {
+      if (!this.hintsFunction) return
       this._codeMirror.showHint({
-        hint: () => {
-          return {
-            list: [{
-              displayText: 'showList',
-              text: JSON.stringify({
-                'cmdType': 'showList',
-                'cmdData': {
-                  'params': [{
-                    'entity': 'TYPE-ENTITY-CODE',
-                    'method': 'select',
-                    'fieldList': ['Dbl-CLICK on left prop panel to add attribute']
-                  }]
-                }
-              }, null, '  ')
-            }, {
-              displayText: 'showForm',
-              text: JSON.stringify({
-                'cmdType': 'showForm',
-                'formCode': 'TYPE HERE A FORM CODE FROM UBM_FORM or remove this line to use a default form for entity',
-                'entity': 'TYPE HERE A ENTITY CODE',
-                'instanceID': 'REPLACE IT by ID value (to edit element) or remove this line'
-              }, null, '  ')
-            }, {
-              displayText: 'showReport',
-              text: JSON.stringify({
-                cmdType: 'showReport',
-                description: 'OPTIONAL report form caption',
-                cmdData: {
-                  reportCode: 'type here report code',
-                  reportType: 'html or pdf',
-                  reportParams: { // if passed report viewer will skip showing parameters enter form to user
-                    paramName: 'param value'
-                  }
-                }
-              }, null, '  ')
-            }],
-            from: this._codeMirror.getCursor(),
-            to: this._codeMirror.getCursor()
-          }
-        }
+        hint: this.hintsFunction
       })
     }
   }