+++ packages/ubm/public/forms/ubm_navshortcut-fm.vue
@@ -128,6 +128,6 @@ module.exports.default = {
       'iconCls'
     ]),
     ...mapGetters(['loading'])
-  },
+  }
 }
 </script>